% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{add_centroid}
\alias{add_centroid}
\title{Calculate the centroid}
\usage{
add_centroid(
  data,
  include_keypoints = NULL,
  exclude_keypoints = NULL,
  centroid_name = "centroid"
)
}
\arguments{
\item{data}{a movement data frame}

\item{include_keypoints}{choose a subset of keypoints used for computing the centorid}

\item{exclude_keypoints}{chosose a subset of keypoints that are excluded when conputing the centroid}
}
\description{
Calculate the centroid
}
