% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{adjust_direction}
\alias{adjust_direction}
\title{Adjust direction
Constrains the direction to be between 0 and 2pi}
\usage{
adjust_direction(direction)
}
\arguments{
\item{direction}{Direction}
}
\description{
Adjust direction
Constrains the direction to be between 0 and 2pi
}
\keyword{internal}
