% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animovement-package.R
\docType{package}
\name{animovement-package}
\alias{animovement}
\alias{animovement-package}
\title{animovement: An R toolbox for analysing animal movement across space and time}
\description{
An R toolbox for analysing animal movement across space and time.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.roald-arboel.com/animovement/}
  \item \url{https://github.com/roaldarbol/animovement/}
  \item Report bugs at \url{https://github.com/roaldarbol/animovement/issues}
}

}
\author{
\strong{Maintainer}: Mikkel Roald-Arbøl \email{trackballr.htmc8@passmail.net} (\href{https://orcid.org/0000-0002-9998-0058}{ORCID})

}
\keyword{internal}
