% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{animovement_conflicts}
\alias{animovement_conflicts}
\title{Conflicts between the animovement and other packages}
\usage{
animovement_conflicts(pkg = animovement_packages())
}
\arguments{
\item{pkg}{character. A vector of packages to check conflicts for. The default is all \emph{animovement} packages.}
}
\value{
An object of class 'animovement_conflicts': A named list of character vectors where the names are the conflicted objects, and the
content are the names of the package namespaces containing the object, in the order they appear on the \code{\link{search}} path.
}
\description{
This function lists all the conflicts among \emph{animovement} packages and between \emph{animovement} packages and other attached packages.
It can also be used to check conflicts for any other attached packages.
}
\examples{
# Check conflicts between animovement packages and all attached packages
animovement_conflicts()

# Check conflicts among all attached packages
animovement_conflicts(sub("package:", "", search()[-1]))
}
\seealso{
\code{\link{animovement}}
}
