% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{animovement_install}
\alias{animovement_install}
\title{Install (missing) animovement packages}
\usage{
animovement_install(
  ...,
  only.missing = TRUE,
  install = TRUE,
  repos = getOption("repos")
)
}
\arguments{
\item{\dots}{comma-separated package names, quoted or unquoted, or vectors of package names. If left empty, all packages returned by \code{\link{animovement_packages}} are checked.}

\item{only.missing}{logical. \code{TRUE} only installs packages that are unavailable. \code{FALSE} installs all packages, even if they are available.}

\item{install}{logical. \code{TRUE} will proceed to install packages, whereas \code{FALSE} (recommended) will print the installation command asking you to run it in a clean R session.}

\item{repos}{the repositories to install from. Defaults to \code{getOption("repos")}. Note that the \emph{animovement} packages are published on R-universe, so this must include \code{https://animovement.r-universe.dev}.}
}
\value{
\code{animovement_install} returns \code{NULL} invisibly.
}
\description{
This function (by default) checks if any \emph{animovement} package is missing and installs the missing package(s).
}
\note{
There is also the possibility to set \code{options(animovement.install = TRUE)} before \code{library(animovement)}, which will call \code{animovement_install()} before loading any packages to make sure all packages are available.
If you are using a \code{.animovement} configuration file inside a project (see vignette), you can also place \code{_opt_animovement.install = TRUE} before the list of packages in that file.
}
\seealso{
\code{\link{animovement_update}}, \code{\link{animovement}}
}
