% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{animovement_packages}
\alias{animovement_packages}
\title{List all packages in the animovement}
\usage{
animovement_packages(extensions = TRUE, include.self = TRUE)
}
\arguments{
\item{extensions}{logical. \code{TRUE} appends the set of core packages with all packages found in \code{options("animovement.extend")}.}

\item{include.self}{logical. Include the \emph{animovement} package in the list?}
}
\value{
A character vector of package names.
}
\description{
Core packages are first fetched from a project-level configuration file (named \code{.animovement}, if found), otherwise the standard set of core packages is returned.
In addition, if \code{extensions = TRUE}, any packages used to extend the \emph{animovement} for the current
session are also returned.
}
\examples{
animovement_packages()
}
\seealso{
\code{\link{animovement_extend}}, \code{\link{animovement}}
}
