% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{arrange.ani_df}
\alias{arrange.ani_df}
\title{Arrange rows of an ani_df}
\usage{
\method{arrange}{ani_df}(.data, ..., .by_group = FALSE)
}
\arguments{
\item{.data}{An ani_df object}

\item{...}{Variables to order by}

\item{.by_group}{If TRUE, arrange within groups}
}
\value{
An arranged ani_df
}
\description{
Arrange rows of an ani_df
}
