% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{as.data.frame.ani_df}
\alias{as.data.frame.ani_df}
\title{Convert ani_df to regular data frame}
\usage{
\method{as.data.frame}{ani_df}(x, ...)
}
\arguments{
\item{x}{An ani_df object}

\item{...}{Additional arguments}
}
\value{
A regular data frame
}
\description{
Convert ani_df to regular data frame
}
