% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_class.R
\name{as_ani_df}
\alias{as_ani_df}
\title{Convert a data frame to ani_df}
\usage{
as_ani_df(data)
}
\arguments{
\item{data}{A data frame with appropriate columns}
}
\value{
An ani_df object
}
\description{
Convert a data frame to ani_df
}
