% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_trackball.R
\name{augment_trackball}
\alias{augment_trackball}
\title{Augment Trackball Data}
\usage{
augment_trackball(
  data,
  x = y_Right,
  y = y_Left,
  time = NULL,
  sampling_rate = 125,
  rollmean_k = 30,
  mouse_dpcm = 394
)
}
\arguments{
\item{data}{Data frame}

\item{x}{Y-data from first sensor.}

\item{y}{Y-data from second sensor.}

\item{time}{Which variable contains time. If no column includes time, use sampling_rate.}

\item{sampling_rate}{Sampling rate of the sensors.}

\item{rollmean_k}{Parameter for smoothing filter.}

\item{mouse_dpcm}{"Dots-per-cm" - or how little movement the mouse can detect.}
}
\value{
An augmented data frame
}
\description{
Augment Trackball Data
}
