% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{calculate_angular_difference}
\alias{calculate_angular_difference}
\title{Calculate angular difference}
\usage{
calculate_angular_difference(from_angle, to_angle)
}
\arguments{
\item{from_angle}{From angle}

\item{to_angle}{To angle}
}
\description{
Calculate angular difference
}
\keyword{internal}
