% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_statistics.R
\name{calculate_circular_mad}
\alias{calculate_circular_mad}
\title{Calculate circular Median Absolute Deviation (MAD)}
\usage{
calculate_circular_mad(angles)
}
\arguments{
\item{angles}{Vector of angles}
}
\description{
Calculate circular Median Absolute Deviation (MAD)
}
\keyword{internal}
