% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_derivative.R
\name{calculate_derivative}
\alias{calculate_derivative}
\title{Calculate the derivative (dx/dt)
Calculate the derivative (dx/dt) with four arguments}
\usage{
calculate_derivative(from_x, to_x, from_t, to_t)
}
\arguments{
\item{from_x}{Current x value}

\item{to_x}{Lagging x value}

\item{from_t}{Current timestamp}

\item{to_t}{Lagging timestamp}
}
\description{
Calculate the derivative (dx/dt)
Calculate the derivative (dx/dt) with four arguments
}
\keyword{internal}
