% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{calculate_direction}
\alias{calculate_direction}
\title{Calculate direction
Calculate direction (angle) from x and y distance using the (two-argument) arc-tangent. Converts to \code{circular}.}
\usage{
calculate_direction(dx, dy)
}
\arguments{
\item{dx}{dx}

\item{dy}{dy}
}
\description{
Calculate direction
Calculate direction (angle) from x and y distance using the (two-argument) arc-tangent. Converts to \code{circular}.
}
\keyword{internal}
