% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate distance (Pythagoras)
Calculate distance from an x and y distance, using Pythagoras theorem.}
\usage{
calculate_distance(dx, dy)
}
\arguments{
\item{dx}{dx}

\item{dy}{dy}
}
\description{
Calculate distance (Pythagoras)
Calculate distance from an x and y distance, using Pythagoras theorem.
}
\keyword{internal}
