% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{calculate_kinematics}
\alias{calculate_kinematics}
\title{Calculate kinematics}
\usage{
calculate_kinematics(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
An data frame data frame with kinematics calculated
}
\description{
Calculate kinematics.
}
