% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_speed.R
\name{calculate_speed}
\alias{calculate_speed}
\title{Calculate Speed from Position Data}
\usage{
calculate_speed(x, y, time)
}
\arguments{
\item{x}{Numeric vector of x coordinates}

\item{y}{Numeric vector of y coordinates}

\item{time}{Numeric vector of time values}
}
\value{
Numeric vector of speeds. The first value will be NA since speed
requires two positions to calculate.
}
\description{
Calculates the instantaneous speed from x, y coordinates and time data.
Speed is computed as the absolute magnitude of velocity (change in position over time).
}
\examples{
\dontrun{
# Inside dplyr pipeline
data |>
  group_by(keypoint) |>
  mutate(speed = calculate_speed(x, y, time))
}

}
