% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_statistics.R
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Calculate summary statistics}
\usage{
calculate_statistics(data, threshold_velocity)
}
\arguments{
\item{data}{A kinematics data frame}

\item{threshold_velocity}{Choose which observations to use based on the velocity. A number, "auto" or "none". Can take a number (e.g. estimated from histograms) or "auto". "auto" fits a density function to the velocities and tries to identify a local minimum between the first and second local maxima, and uses that as the threshold. "none" keeps all observations.}
}
\value{
An data frame data frame with kinematics calculated
}
\description{
Calculate summary statistics for tracks
}
