% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_straightness.R
\name{calculate_straightness}
\alias{calculate_straightness}
\title{Calculate straightness measures}
\usage{
calculate_straightness(data, straightness = c("A", "B", "C", "D"))
}
\arguments{
\item{data}{Data frame}

\item{straightness}{hich method to calculate path straightness. Choose between "A" (default), "B", "C"... or a combination (e.g. "c("A","B")"). See description for details about the different calculations.}
}
\description{
Calculate straightness measures
}
\keyword{internal}
