% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{$.ani_df}
\alias{$.ani_df}
\title{Extract column from ani_df with $}
\usage{
\method{$}{ani_df}(x, name)
}
\arguments{
\item{x}{An ani_df object}

\item{name}{Column name}
}
\value{
A vector
}
\description{
Extract column from ani_df with $
}
