% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_na_timing.R
\name{check_na_timing}
\alias{check_na_timing}
\title{Visualize the timing of missing values in the data}
\usage{
check_na_timing(data, by_keypoint = TRUE)
}
\arguments{
\item{data}{A data frame containing at least the columns \code{x}, \code{individual},
and \code{keypoint}.}

\item{by_keypoint}{Logical. If \code{TRUE}, generates a separate plot for each keypoint.
If \code{FALSE}, creates a single aggregated plot for all keypoints. Default is \code{TRUE}.}
}
\value{
A \code{patchwork} object combining one or more ggplots that show the
timing of missing values (\code{NA}) in the data.
}
\description{
This function generates a plot to visualize where missing values (\code{NA}s) occur
in the data over time. It can display separate plots for each keypoint or a
single aggregated plot for all keypoints.
}
\details{
\itemize{
\item Missing values are highlighted using a red (\code{indianred2}) color, and non-missing
values are shown in blue (\code{steelblue}).
\item The function uses the \code{patchwork} package to combine multiple plots when \code{by_keypoint = TRUE}.
}
}
\examples{
library(dplyr)
library(ggplot2)
library(patchwork)
data <- dplyr::tibble(
  x = c(1, 2, NA, 4, NA, 6),
  individual = rep("A", 6),
  keypoint = factor(rep(c("head", "arm"), each = 3))
)
check_na_timing(data, by_keypoint = TRUE)

}
