% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_kinematics.R
\name{clean_kinematics}
\alias{clean_kinematics}
\title{Clean kinematics}
\usage{
clean_kinematics(data)
}
\arguments{
\item{data}{tidy movement data frame with kinematics}
}
\value{
a clean kinematics data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
