% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_kinematics.R
\name{compute_kinematics}
\alias{compute_kinematics}
\title{Compute kinematics}
\usage{
compute_kinematics(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
An data frame data frame with kinematics computed
}
\description{
Compute kinematics.
}
