% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_statistics.R
\name{compute_statistics}
\alias{compute_statistics}
\title{Compute summary statistics}
\usage{
compute_statistics(data, threshold_velocity)
}
\arguments{
\item{data}{A kinematics data frame}

\item{threshold_velocity}{Choose which observations to use based on the velocity. A number, "auto" or "none". Can take a number (e.g. estimated from histograms) or "auto". "auto" fits a density function to the velocities and tries to identify a local minimum between the first and second local maxima, and uses that as the threshold. "none" keeps all observations.}
}
\value{
An data frame data frame with kinematics computed
}
\description{
Compute summary statistics for tracks
}
