% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_kinematics.R
\name{ensure_circular}
\alias{ensure_circular}
\title{Ensure that values are circular}
\usage{
ensure_circular(x)
}
\arguments{
\item{x}{Vector of values}
}
\description{
Ensure that values are circular
}
\keyword{internal}
