% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_deeplabcut.R
\name{ensure_dlc_expected_header_levels}
\alias{ensure_dlc_expected_header_levels}
\title{Ensure that the .csv file contains the expected index column levels.}
\usage{
ensure_dlc_expected_header_levels(path)
}
\arguments{
\item{path}{Path to the .csv file.}
}
\description{
Ensure that the .csv file contains the expected index column levels.
}
\keyword{internal}
