% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_files.R
\name{ensure_file_exists_when_expected}
\alias{ensure_file_exists_when_expected}
\title{Ensure that the file exists (or not) as needed.}
\usage{
ensure_file_exists_when_expected(path, expected_permission)
}
\arguments{
\item{path}{Path(s) to the file.}

\item{expected_permission}{Expected access permission(s) for the file. If
"r", the file is expected to be readable. If "w", the file is expected to
be writable. If "rw", the file is expected to be both readable and
writable. Default: "r".}
}
\description{
Ensure that the file exists (or not) as needed. This depends on the expected usage (read and/or write).
}
\keyword{internal}
