% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_files.R
\name{ensure_file_has_access_permissions}
\alias{ensure_file_has_access_permissions}
\title{Ensure that the file has the expected access permission(s).}
\usage{
ensure_file_has_access_permissions(path, expected_permission)
}
\arguments{
\item{path}{Path(s) to the file.}

\item{expected_permission}{Expected access permission(s) for the file. If
"r", the file is expected to be readable. If "w", the file is expected to
be writable. If "rw", the file is expected to be both readable and
writable. Default: "r".}
}
\description{
Ensure that the file has the expected access permission(s).
}
