% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_files.R
\name{ensure_file_has_expected_suffix}
\alias{ensure_file_has_expected_suffix}
\title{Ensure that the file has one of the expected suffix(es).}
\usage{
ensure_file_has_expected_suffix(path, expected_suffix)
}
\arguments{
\item{path}{Path(s) to the file.}

\item{expected_suffix}{Expected suffix(es) for the file. If NULL (default), this check is skipped.}
}
\description{
Ensure that the file has one of the expected suffix(es).
}
\keyword{internal}
