% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_files.R
\name{ensure_file_has_headers}
\alias{ensure_file_has_headers}
\title{Ensure file has headers}
\usage{
ensure_file_has_headers(path)
}
\arguments{
\item{path}{Path(s) to the file.}
}
\description{
Ensure file has headers
}
\keyword{internal}
