% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_files.R
\name{ensure_is_not_dir}
\alias{ensure_is_not_dir}
\title{Ensure that the path does not point to a directory.}
\usage{
ensure_is_not_dir(path)
}
\arguments{
\item{path}{Path(s) to the file.}
}
\description{
Ensure that the path does not point to a directory.
}
\keyword{internal}
