% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_trackball.R
\name{ensure_trackball_setup}
\alias{ensure_trackball_setup}
\title{Ensure that a valid setup is given}
\usage{
ensure_trackball_setup(setup)
}
\arguments{
\item{setup}{Experimental setup used. Expects either "of_free", "of_fixed" or "fictrac".}
}
\description{
Ensure that a valid setup is given
}
\keyword{internal}
