% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_class.R
\name{example_tbl}
\alias{example_tbl}
\title{Create example ani_df data}
\usage{
example_tbl(class = NULL)
}
\arguments{
\item{class}{Unused parameter (for future extension)}
}
\value{
An example ani_df object
}
\description{
Create example ani_df data
}
\examples{
example_tbl()
}
