% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{filter.ani_df}
\alias{filter.ani_df}
\title{Filter rows of an ani_df}
\usage{
\method{filter}{ani_df}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{An ani_df object}

\item{...}{Logical predicates}

\item{.preserve}{Keep group structure}
}
\value{
A filtered ani_df
}
\description{
Filter rows of an ani_df
}
