% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rollmedian.R
\name{filter_rollmedian}
\alias{filter_rollmedian}
\title{Apply Rolling Median Filter}
\usage{
filter_rollmedian(x, window_width = 5, min_obs = 1, ...)
}
\arguments{
\item{x}{Numeric vector to filter}

\item{window_width}{Integer specifying window size for rolling calculation}

\item{min_obs}{Minimum number of non-NA values required (default: 1)}

\item{...}{Additional parameters to be passed to \code{roll::roll_median}}
}
\value{
Filtered numeric vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Applies a rolling median filter to a numeric vector using the roll package.
}
