% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get/extract metadata}
\usage{
get_metadata(data)
}
\arguments{
\item{data}{movement data frame}
}
\value{
the metadata associated with the movement data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
