% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{group_by.ani_df}
\alias{group_by.ani_df}
\title{Group an ani_df}
\usage{
\method{group_by}{ani_df}(.data, ..., .add = FALSE, .drop = TRUE)
}
\arguments{
\item{.data}{An ani_df object}

\item{...}{Variables to group by}

\item{.add}{If TRUE, add to existing groups}

\item{.drop}{Drop unused factor levels}
}
\value{
A grouped ani_df
}
\description{
Group an ani_df
}
