% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_metadata.R
\name{init_metadata}
\alias{init_metadata}
\title{Initiate movement metadata}
\usage{
init_metadata(data, metadata = list())
}
\arguments{
\item{data}{movement data frame}
}
\value{
data frame with metadata
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
