% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_class.R
\name{is_ani_df}
\alias{is_ani_df}
\title{Check if object is an ani_df}
\usage{
is_ani_df(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical: TRUE if x inherits from ani_df
}
\description{
Check if object is an ani_df
}
