% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trackball.R
\name{join_trackball_files}
\alias{join_trackball_files}
\title{Join data files with non-matching time stamps}
\usage{
join_trackball_files(data_list, sampling_rate)
}
\arguments{
\item{data_list}{List of 2 dataframes}

\item{sampling_rate}{Sampling rate tells the function how long time it should integrate over. A sampling rate of 60(Hz) will mean windows of 1/60 sec are used to integrate over.}
}
\description{
Join data files with non-matching time stamps
}
