% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_transform.R
\name{map_to_cartesian}
\alias{map_to_cartesian}
\title{Map from polar to Cartesian coordinates}
\usage{
map_to_cartesian(data)
}
\arguments{
\item{data}{movement data frame with polar coordinates}
}
\value{
movement data frame with Cartesian coordinates
}
\description{
Map from polar to Cartesian coordinates
}
