% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_transform.R
\name{map_to_polar}
\alias{map_to_polar}
\title{Map from Cartesian to polar coordinates}
\usage{
map_to_polar(data)
}
\arguments{
\item{data}{movement data frame with Cartesian coordinates}
}
\value{
movement data frame with polar coordinates
}
\description{
Map from Cartesian to polar coordinates
}
