% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
A vector of the most frequently occuring value(s)
}
\description{
In statistics, the mode is the value that appears most often in a set of data values.
}
\examples{
my_vector <- c(1,2,3,3,4,5)
mode(my_vector)
}
