% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{mutate.ani_df}
\alias{mutate.ani_df}
\title{Mutate columns in an ani_df}
\usage{
\method{mutate}{ani_df}(.data, ...)
}
\arguments{
\item{.data}{An ani_df object}

\item{...}{Name-value pairs of expressions}
}
\value{
An ani_df with modified columns
}
\description{
Mutate columns in an ani_df
}
