% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{names<-.ani_df}
\alias{names<-.ani_df}
\title{Rename columns with names<-}
\usage{
\method{names}{ani_df}(x) <- value
}
\arguments{
\item{x}{An ani_df object}

\item{value}{New column names}
}
\value{
Modified ani_df
}
\description{
Rename columns with names<-
}
