% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_class.R
\name{new_ani_df}
\alias{new_ani_df}
\title{Create a new ani_df object (internal constructor)}
\usage{
new_ani_df(x, metadata = list())
}
\arguments{
\item{x}{A data frame to convert to ani_df}

\item{metadata}{Optional metadata list}
}
\value{
An ani_df object
}
\description{
Create a new ani_df object (internal constructor)
}
\keyword{internal}
