% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bonsai.R
\name{read_bonsai}
\alias{read_bonsai}
\title{Read centroid tracking data from Bonsai}
\usage{
read_bonsai(path)
}
\arguments{
\item{path}{Path to a Bonsai data file}
}
\value{
a movement dataframe
}
\description{
Read centroid tracking data from Bonsai
}
