% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idtracker.R
\name{read_idtracker}
\alias{read_idtracker}
\title{Read idtracker.ai data}
\usage{
read_idtracker(data)
}
\arguments{
\item{data}{An idtracker.ai data frame}
}
\value{
a movement dataframe
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
