% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sleap.R
\name{read_sleap}
\alias{read_sleap}
\title{Read SLEAP data}
\usage{
read_sleap(path)
}
\arguments{
\item{path}{A SLEAP analysis data frame in HDF5 (.h5) format}
}
\value{
a movement dataframe
}
\description{
Read SLEAP data
}
