% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_trackball}
\alias{read_trackball}
\title{Read Trackball Data}
\usage{
read_trackball(
  filepaths,
  configuration = c("free", "fixed"),
  time_col,
  sampling_rate,
  mouse_dpcm = NULL
)
}
\arguments{
\item{filepaths}{To file paths, one for each sensor}

\item{configuration}{Open- or closed-loop configuration}

\item{time_col}{Which column contains the information about time. Can be specified either by the column number (numeric) or the name of the column if it has one (character). Should either be a datetime (POSIXt) or seconds (numeric).}

\item{sampling_rate}{Sampling rate tells the function how long time it should intergrate over. A sampling rate of 60(Hz) will mean windows of 1/60 sec are used to integrate over.}

\item{mouse_dpcm}{If using computer mice, you might be getting unit-less data out. However, computer mice have a factor called "dots-per-cm", which you can use to convert your estimates into centimeters.}
}
\value{
A list of data frames
}
\description{
Read Trackball Data
}
