% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_trackball_data}
\alias{read_trackball_data}
\title{Read Trackball Data}
\usage{
read_trackball_data(
  folder_path,
  sensor_names = c("Left", "Right"),
  format = "RDS"
)
}
\arguments{
\item{folder_path}{Path to folders containing data files.}

\item{sensor_names}{Names of the two sensors.}

\item{format}{Format of data files (default: RDS)}
}
\value{
A list of data frames
}
\description{
Read Trackball Data
}
