% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trex.R
\name{read_trex}
\alias{read_trex}
\title{Read TRex data}
\usage{
read_trex(path)
}
\arguments{
\item{path}{Path to a TRex data frame in CSV format.}
}
\value{
a movement dataframe
}
\description{
Read TRex data
}
