% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{relocate.ani_df}
\alias{relocate.ani_df}
\title{Relocate columns in an ani_df}
\usage{
\method{relocate}{ani_df}(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{An ani_df object}

\item{...}{Columns to relocate}

\item{.before}{Column to place before}

\item{.after}{Column to place after}
}
\value{
An ani_df with relocated columns
}
\description{
Relocate columns in an ani_df
}
