% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{rename.ani_df}
\alias{rename.ani_df}
\title{Rename columns in an ani_df}
\usage{
\method{rename}{ani_df}(.data, ...)
}
\arguments{
\item{.data}{An ani_df object}

\item{...}{Name-value pairs for renaming}
}
\value{
An ani_df with renamed columns
}
\description{
Rename columns in an ani_df
}
