% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{select.ani_df}
\alias{select.ani_df}
\title{Select columns from an ani_df}
\usage{
\method{select}{ani_df}(.data, ...)
}
\arguments{
\item{.data}{An ani_df object}

\item{...}{Columns to select}
}
\value{
An ani_df with selected columns
}
\description{
Select columns from an ani_df
}
