% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_metadata.R
\name{set_start_datetime}
\alias{set_start_datetime}
\title{Set starting datetime}
\usage{
set_start_datetime(data, start_datetime)
}
\arguments{
\item{data}{movement data frame}

\item{start_datetime}{starting datetime. provided either as POSIXt, or as a string that can be parsed by the \emph{anytime} package.}
}
\value{
movement data frame with starting datetime in metadata
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
