% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_metadata.R
\name{set_uuid}
\alias{set_uuid}
\title{Set UUID}
\usage{
set_uuid(data, length = 20)
}
\arguments{
\item{data}{movement data frame}

\item{length}{length of identifier. (default: 20)}
}
\value{
data frame with the "uuid" metadata field filled out
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Adds a unique identifier (UUID) to the data frames metadata
}
