% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{slice.ani_df}
\alias{slice.ani_df}
\title{Slice rows from an ani_df}
\usage{
\method{slice}{ani_df}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{An ani_df object}

\item{...}{Integer row positions}

\item{.preserve}{Keep group structure}
}
\value{
A sliced ani_df
}
\description{
Slice rows from an ani_df
}
