% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_savgol.R
\name{smooth_by_savgol}
\alias{smooth_by_savgol}
\title{Smooth a trajectory using a Savitzky-Golay filter}
\usage{
smooth_by_savgol(data, p = 3, n = p + 3 - p\%\%2, ...)
}
\arguments{
\item{data}{a movement data frame}

\item{p}{polynomial order (passed to \code{\link[signal]{sgolayfilt}}).}

\item{n}{Filter length (or window size), must be an odd number.  Passed to
\code{\link[signal]{sgolayfilt}}.}

\item{...}{Additional arguments are passed to
\code{\link[signal]{sgolayfilt}}.}
}
\value{
A new trajectory which is a smoothed version of the input trajectory.
}
\description{
Smooths a trajectory using a Savitzky-Golay smoothing filter.
}
\details{
Consider carefully the effects of smoothing a trajectory with temporal gaps
in the data. If the smoothed trajectory is used to derive speed and/or
acceleration, it may be advisable to fill in the gaps before smoothing,
possibly by calling \code{TrajResampleTime}.
}
\examples{
set.seed(3)
trj <- TrajGenerate(500, random = TRUE, angularErrorSd = .25)
smoothed <- TrajSmoothSG(trj, 3, 31)
plot(trj)
plot(smoothed, col = "red", add = TRUE)

}
\seealso{
\code{\link[signal]{sgolayfilt}}
}
