% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_movement.R
\name{smooth_derivatives}
\alias{smooth_derivatives}
\title{Smooth derivatives (when original data measured differences between movements)}
\usage{
smooth_derivatives(data, method, window_width)
}
\arguments{
\item{data}{Data frame}

\item{method}{Which smoothing method to use. options:  "rolling_median (default), "rolling_mean".}

\item{window_width}{How many observations to use for rolling window filters (e.g. "rolling_mean" or "rolling_median").}
}
\description{
Smooth derivatives (when original data measured differences between movements)
}
\keyword{internal}
