% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_movement.R
\name{smooth_movement}
\alias{smooth_movement}
\title{Smooth movement}
\usage{
smooth_movement(
  data,
  method = c("rolling_median"),
  window_width = 5,
  min_obs = 1,
  use_derivatives = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{method}{Which smoothing method to use. options:  "rolling_median (default), "rolling_mean".}

\item{window_width}{How many observations to use for rolling window filters (e.g. "rolling_mean" or "rolling_median").}

\item{use_derivatives}{whether to use the derivatives (difference between frames) to perform the smoothing. Useful for trackball data.}
}
\value{
A movement data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Filtering/smoothing tracks is standard practice to root out noise in movement data.
Here we provide some filter functions to do this. The function expects the data to be in the standard format,
containing at least x, y and time variables.
}
