% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_movement.R
\name{smooth_track}
\alias{smooth_track}
\title{Smooth tracks}
\usage{
smooth_track(data, method = c("rolling_median"), window_width = 5)
}
\arguments{
\item{data}{Data frame}

\item{method}{Which smoothing method to use. options:  "rolling_median (default), "rolling_mean".}

\item{window_width}{How many observations to use for rolling window filters (e.g. "rolling_mean" or "rolling_median").}
}
\value{
A movement data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Filtering/smoothing tracks is standard practice to root out noise in movement data.
Here we provide some filter functions to do this. The function expects the data to be in the standard format,
containing at least x, y and time variables.
}
