% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{[.ani_df}
\alias{[.ani_df}
\title{Subset ani_df with [}
\usage{
\method{[}{ani_df}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{An ani_df object}

\item{i}{Row indices}

\item{j}{Column indices}

\item{...}{Additional arguments}

\item{drop}{If TRUE, simplify to vector when possible}
}
\value{
A subset ani_df
}
\description{
Subset ani_df with [
}
