% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{[[.ani_df}
\alias{[[.ani_df}
\title{Extract single column from ani_df with [[}
\usage{
\method{[[}{ani_df}(x, i, ...)
}
\arguments{
\item{x}{An ani_df object}

\item{i}{Column index or name}

\item{...}{Additional arguments}
}
\value{
A vector or data frame
}
\description{
Extract single column from ani_df with [[
}
