% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{[[<-.ani_df}
\alias{[[<-.ani_df}
\title{Column assignment for ani_df with [[<-}
\usage{
\method{[[}{ani_df}(x, i, ...) <- value
}
\arguments{
\item{x}{An ani_df object}

\item{i}{Column index or name}

\item{...}{Additional arguments}

\item{value}{Replacement value}
}
\value{
Modified ani_df
}
\description{
Column assignment for ani_df with [[<-
}
