% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{[<-.ani_df}
\alias{[<-.ani_df}
\title{Subset assignment for ani_df with [<-}
\usage{
\method{[}{ani_df}(x, i, j, ...) <- value
}
\arguments{
\item{x}{An ani_df object}

\item{i}{Row indices}

\item{j}{Column indices}

\item{...}{Additional arguments}

\item{value}{Replacement values}
}
\value{
Modified ani_df
}
\description{
Subset assignment for ani_df with [<-
}
