% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_class.R
\name{tbl_sum.ani_df}
\alias{tbl_sum.ani_df}
\title{Custom tibble summary for ani_df}
\usage{
\method{tbl_sum}{ani_df}(x, ...)
}
\arguments{
\item{x}{An ani_df object}

\item{...}{Additional arguments (unused)}
}
\value{
Named character vector with summary information
}
\description{
Custom tibble summary for ani_df
}
