% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackballr-package.R
\docType{package}
\name{trackballr-package}
\alias{trackballr}
\alias{trackballr-package}
\title{trackballr: Speed Up Trackball Analysis}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
