% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_transform.R
\name{translate_coords_keypoint}
\alias{translate_coords_keypoint}
\title{Translate coordinates relative to a keypoint}
\usage{
translate_coords_keypoint(data, to_keypoint)
}
\arguments{
\item{data}{movement data frame}

\item{to_keypoint}{all other coordinates becomes relative to this keypoint.}
}
\description{
Translate coordinates relative to a keypoint
}
\keyword{internal}
