% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_df_methods.R
\name{ungroup.ani_df}
\alias{ungroup.ani_df}
\title{Ungroup an ani_df}
\usage{
\method{ungroup}{ani_df}(x, quiet = FALSE, ...)
}
\arguments{
\item{x}{An ani_df object}

\item{quiet}{If FALSE, shows a warning about ungrouping}

\item{...}{Additional arguments passed to dplyr::ungroup}
}
\value{
An ungrouped ani_df
}
\description{
Ungroup an ani_df
}
