% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_animalta.R
\name{validate_animalta}
\alias{validate_animalta}
\title{#' Validate AnimalTA .csv files.}
\usage{
validate_animalta(path)
}
\arguments{
\item{path}{An AnimalTA data frame}
}
\description{
A short description...
}
\keyword{internal}
