% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_deeplabcut.R
\name{validate_deeplabcut_csv}
\alias{validate_deeplabcut_csv}
\title{Validate DeepLabCut-style .csv files.}
\usage{
validate_deeplabcut_csv(path)
}
\arguments{
\item{path}{Path to the .csv file.}
}
\description{
The validator ensures that the file contains the expected index column levels.
}
\keyword{internal}
