% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_files.R
\name{validate_files}
\alias{validate_files}
\title{Validate file}
\usage{
validate_files(paths, expected_permission = "r", expected_suffix = NULL)
}
\arguments{
\item{expected_permission}{Expected access permission(s) for the file. If
"r", the file is expected to be readable. If "w", the file is expected to
be writable. If "rw", the file is expected to be both readable and
writable. Default: "r".}

\item{expected_suffix}{Expected suffix(es) for the file. If NULL (default), this check is skipped.}

\item{path}{Path to the file.}
}
\description{
The validator ensures that the file:
\itemize{
\item is not a directory,
\item exists if it is meant to be read,
\item does not exist if it is meant to be written,
\item has one of the expected suffix(es).
}
}
