% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_time.R
\name{validate_time}
\alias{validate_time}
\title{Title}
\usage{
validate_time(col_time, sampling_rate)
}
\arguments{
\item{col_time}{Column for time.}

\item{sampling_rate}{Sampling rate.}
}
\description{
Title
}
\keyword{internal}
