% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animovement.R
\docType{package}
\name{animovement}
\alias{animovement-package}
\alias{animovement}
\title{The animovement}
\description{
The \emph{animovement} is an extensible suite of R packages. It is a descendant of the \href{https://fastverse.github.io/fastverse/}{fastverse}.
}
\section{Functions in the \emph{animovement} Package}{


Functions to extend or reduce the number of packages in the \emph{animovement}

\code{\link[=animovement_extend]{animovement_extend()}}\cr
\code{\link[=animovement_detach]{animovement_detach()}}

Function to display conflicts for \emph{animovement} packages (or any other attached packages)

\code{\link[=animovement_conflicts]{animovement_conflicts()}}

Function to update \emph{animovement} packages (and dependencies) and install (missing) packages

\code{\link[=animovement_update]{animovement_update()}}\cr
\code{\link[=animovement_install]{animovement_install()}}

Utilities to retrieve the names of \emph{animovement} packages (and dependencies), their update status and produce a situation report.

\code{\link[=animovement_packages]{animovement_packages()}}\cr
\code{\link[=animovement_deps]{animovement_deps()}}\cr
\code{\link[=animovement_sitrep]{animovement_sitrep()}}
}

\section{\emph{animovement} Options}{

\itemize{
\item \code{options(animovement.quiet = TRUE)} will disable all automatic messages (including conflict reporting) when calling \code{library(animovement)}, \code{\link{animovement_extend}}, \code{\link[=animovement_update]{animovement_update(install = TRUE)}} and \code{\link{animovement_install}}.
\item \code{options(animovement.styling = FALSE)} will disable all styling applied to text printed to the console.
\item \code{options(animovement.extend = c(...))} can be set before calling \code{library(animovement)} to extend the animovement with some packages for the session. The same can be done with the
\code{\link{animovement_extend}} function after \code{library(animovement)}, which will also populate \code{options("animovement.extend")}.
\item \code{options(animovement.install = TRUE)} can be set before \code{library(animovement)} to install any missing packages beforehand. See also \code{\link{animovement_install}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://animovement.dev}
  \item \url{https://animovement.dev/animovement}
  \item \url{https://github.com/animovement/animovement}
  \item Report bugs at \url{https://github.com/animovement/animovement/issues}
}

}
\author{
\strong{Maintainer}: Mikkel Roald-Arbøl \email{animovement.84w1m@passmail.com} (\href{https://orcid.org/0000-0002-9998-0058}{ORCID})

Authors:
\itemize{
  \item Mikkel Roald-Arbøl \email{animovement.84w1m@passmail.com} (\href{https://orcid.org/0000-0002-9998-0058}{ORCID})
}

Other contributors:
\itemize{
  \item Sebastian Krantz (Author of the fastverse, from which the package management code is adapted) [contributor]
  \item Hadley Wickham (Author of the tidyverse, from which that code originally derives) [contributor]
}

}
\keyword{internal}
