% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{animovement_detach}
\alias{animovement_detach}
\title{Detach animovement packages}
\usage{
animovement_detach(
  ...,
  unload = FALSE,
  force = FALSE,
  include.self = TRUE,
  session = FALSE
)
}
\arguments{
\item{\dots}{comma-separated package names, quoted or unquoted, or vectors of package names. If left empty, all packages returned by \code{\link{animovement_packages}} are detached.}

\item{unload}{logical. \code{TRUE} also unloads the packages using \code{\link[=detach]{detach(name, unload = TRUE)}}.}

\item{force}{logical. should a \emph{animovement} package be detached / unloaded even though other attached packages depend on it?}

\item{include.self}{logical. \code{TRUE} also includes the \code{animovement} package - only applicable if \code{\dots} is left empty.}

\item{session}{logical. \code{TRUE} also removes the packages from \code{options("animovement.extend")}, so they will not be attached again with \code{library(animovement)} in the current session. If \code{\dots} is left empty and \code{include.self = TRUE}, this will clear \bold{all} \emph{animovement} options set for the session.}
}
\value{
\code{animovement_detach} returns \code{NULL} invisibly.
}
\description{
Detaches \emph{animovement} packages, removing them from the \code{\link{search}} path.
}
\seealso{
\code{\link{animovement_extend}}, \code{\link{animovement}}
}
