% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{animovement_extend}
\alias{animovement_extend}
\title{Extend the animovement}
\usage{
animovement_extend(
  ...,
  install = FALSE,
  check.conflicts = !isTRUE(getOption("animovement.quiet"))
)
}
\arguments{
\item{\dots}{comma-separated package names, quoted or unquoted, or vectors of package names.}

\item{install}{logical. Install packages not available?}

\item{check.conflicts}{logical. Should conflicts between extension packages and attached packages be checked?}
}
\value{
\code{animovement_extend} returns \code{NULL} invisibly.
}
\description{
Loads additional packages as part of the \emph{animovement}.
}
\details{
When the \emph{animovement} is extended calling \code{animovement_extend(...)}, the packages that are not attached are attached, but conflicts are checked for all specified packages.
An \code{options("animovement.extend")} is set which stores these extension packages, regardless of whether they were already attached or not. When calling
\code{\link{animovement_packages}}, \code{\link{animovement_deps}}, \code{\link{animovement_conflicts}}, \code{\link{animovement_update}}, \code{\link{animovement_sitrep}} or \code{\link{animovement_detach}}, these packages are included as part of the \emph{animovement}.
To extend the \emph{animovement} for the current session when it is not yet loaded, users can also set \code{options(animovement.extend = c(...))}, where \code{c(...)}
is a character vector of package names, before calling \code{library(animovement)}.
}
\seealso{
\code{\link{animovement_detach}}, \code{\link{animovement}}
}
