% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_suggested.R
\name{animovement_install_suggested}
\alias{animovement_install_suggested}
\alias{animovement_show_suggested}
\title{Download all suggested packages}
\usage{
animovement_install_suggested(package = "animovement")

animovement_show_suggested(package = "animovement")
}
\arguments{
\item{package}{Character string specifying the package name. Currently only
\code{"animovement"} is supported.}
}
\value{
Invisible \code{NULL}. Used for side-effect of installing packages.
}
\description{
In \code{animovement}, we have a minimal dependency policy to keep the package
light and fast to install. However, we rely on several packages for testing
and specific features. These "soft dependencies" can be downloaded at once
using this function, allowing you to fully utilize all of animovement's
functionalities without errors.
}
\details{
To reduce the dependency load, animovement by default will not
download all internally needed packages. It will ask the user to download
them only if they are needed. The current function can help install all
packages that animovement and its ecosystem packages might need.
\code{animovement_show_suggested()} is a convenient helper to show the current
list of suggested packages.

This function will check the \code{Suggests} field of animovement and all its
imported packages (aniframe, aniread, aniprocess, animetric, anicheck, anivis),
excluding packages only needed for development or documentation workflows
(knitr, rmarkdown, testthat, pak, here, covr, pkgdown).

Under WebR, packages are installed with \code{webr::install()}, since
\code{utils::install.packages()} cannot build Emscripten packages in the browser.
Otherwise, if \code{{pak}} is installed then \code{pak::pkg_install()} is used, and
failing that \code{utils::install.packages()}. Repositories searched are CRAN, the
animovement R-universe and the Bioconductor R-universe -- the last of which
is where \code{rhdf5} comes from.
}
\examples{
# download all suggested packages
if (FALSE) {
  animovement_install_suggested("animovement")
}

# listing all soft/weak dependencies
animovement_show_suggested()

}
