% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{animovement_sitrep}
\alias{animovement_sitrep}
\title{Get a situation report on the animovement}
\usage{
animovement_sitrep(...)
}
\arguments{
\item{\dots}{arguments other than \code{pkg} passed to \code{\link{animovement_deps}}.}
}
\value{
\code{animovement_sitrep} returns \code{NULL} invisibly.
}
\description{
This function gives a quick overview of the version of R and all
\emph{animovement} packages (including availability updates for packages) and indicates
whether any project-level configuration files are used.
}
\seealso{
\code{\link{animovement_deps}}, \code{\link{animovement}}
}
