% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{animovement_update}
\alias{animovement_update}
\title{Update animovement packages}
\usage{
animovement_update(..., install = FALSE)
}
\arguments{
\item{\dots}{arguments passed to \code{\link{animovement_deps}}.}

\item{install}{logical. \code{TRUE} will proceed to install outdated packages, whereas \code{FALSE} (recommended) will print the installation command asking you to run it in a clean R session.}
}
\value{
\code{animovement_update} returns \code{NULL} invisibly.
}
\description{
This will check all \emph{animovement} packages (and their
dependencies) for updates and (optionally) install those updates.
}
\seealso{
\code{\link{animovement_deps}}, \code{\link{animovement}}
}
