% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kinematics.R
\name{calculate_kinematics}
\alias{calculate_kinematics}
\title{Calculate kinematics}
\usage{
calculate_kinematics(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
A data frame with kinematics calculated
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate kinematics.
}
