% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_statistics.R
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Calculate summary statistics}
\usage{
calculate_statistics(
  data,
  measures = "median_mad",
  straightness = c("A", "B", "C", "D")
)
}
\arguments{
\item{data}{A kinematics data frame}

\item{measures}{Measures of central tendency and dispersion. Options are \code{median_mad} (default) and \code{mean_sd}. See description for more information.}

\item{straightness}{Which method to calculate path straightness. Choose between "A" (default), "B", "C"... or a combination (e.g. "c("A","B")"). See description for details about the different calculations.}
}
\value{
An data frame data frame with kinematics calculated
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate summary statistics for tracks
}
