% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_animalta.R
\name{read_animalta}
\alias{read_animalta}
\title{Read AnimalTA data}
\usage{
read_animalta(path)
}
\arguments{
\item{path}{An AnimalTA data frame}
}
\value{
a movement dataframe
}
\description{
Read AnimalTA data
}
