% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_animalta.R
\name{read_animalta}
\alias{read_animalta}
\title{Read AnimalTA data}
\usage{
read_animalta(path, with_roi = FALSE)
}
\arguments{
\item{path}{An AnimalTA data frame}

\item{with_roi}{Were one or more ROIs used?}
}
\value{
a movement dataframe
}
\description{
Read a data frame from AnimalTA
}
\references{
\itemize{
\item Chiara, V., & Kim, S.-Y. (2023). AnimalTA: A highly flexible and easy-to-use
program for tracking and analysing animal movement in different environments.
\emph{Methods in Ecology and Evolution}, 14, 1699–1707. \doi{0.1111/2041-210X.14115}.
}
}
