% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_animalta.R
\name{read_animalta}
\alias{read_animalta}
\title{Read AnimalTA data}
\usage{
read_animalta(path, with_roi = FALSE)
}
\arguments{
\item{path}{An AnimalTA data frame}

\item{with_roi}{Were one or more ROIs used?}
}
\value{
a movement dataframe
}
\description{
Read AnimalTA data
}
