% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lightningpose.R
\name{read_lightningpose}
\alias{read_lightningpose}
\title{Read LightningPose data}
\usage{
read_lightningpose(path)
}
\arguments{
\item{path}{Path to a LightningPose data file}
}
\value{
a movement dataframe
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Read csv files from LightningPose (LP).
}
