% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_movement.R
\name{read_movement}
\alias{read_movement}
\title{Read movement data}
\usage{
read_movement(data)
}
\arguments{
\item{data}{A movement data frame}
}
\value{
a movement dataframe
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
