% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trackball.R
\name{read_opticalflow}
\alias{read_opticalflow}
\title{Read optical flow sensor file}
\usage{
read_opticalflow(path, col_time, verbose = FALSE)
}
\arguments{
\item{path}{Path to the file.}

\item{col_time}{Which column contains the information about time. Can be specified either by the column number (numeric) or the name of the column if it has one (character). Should either be a datetime (POSIXt) or seconds (numeric).}

\item{verbose}{If \code{FALSE} (default), suppress most warning messages.}
}
\description{
Read optical flow sensor data.
}
\keyword{internal}
