% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sleap.R
\name{read_sleap}
\alias{read_sleap}
\title{Read SLEAP data}
\usage{
read_sleap(path)
}
\arguments{
\item{path}{A SLEAP data frame}
}
\value{
a movement dataframe
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
