% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_trackball_data}
\alias{read_trackball_data}
\title{Read Trackball Data}
\usage{
read_trackball_data(filepaths, configuration = c("free", "fixed"))
}
\arguments{
\item{filepaths}{To file paths, one for each sensor}

\item{configuration}{Open- or closed-loop configuration}
}
\value{
A list of data frames
}
\description{
Read Trackball Data
}
