% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_treadmill.R
\name{read_treadmill}
\alias{read_treadmill}
\title{Read treadmill data}
\usage{
read_treadmill(data)
}
\arguments{
\item{data}{A treadmill data frame}
}
\value{
a movement dataframe
}
\description{
Read treadmill data
}
