% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackballr-package.R
\docType{package}
\name{trackballr-package}
\alias{trackballr}
\alias{trackballr-package}
\title{trackballr: R package for reading trackball data into a standard format}
\description{
R package for reading trackball data into a standard format.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.roald-arboel.com/trackballr/}
}

}
\author{
\strong{Maintainer}: Mikkel Roald-Arbøl \email{trackballr.htmc8@passmail.net} (\href{https://orcid.org/0000-0002-9998-0058}{ORCID})

}
\keyword{internal}
